+++ src/me/anno/fonts/signeddistfields/structs/FloatPtr.kt
@@ -1,3 +1,3 @@
 package me.anno.fonts.signeddistfields.structs
 
-data class DoublePtr(var value: Double = 0.0)
+data class FloatPtr(var value: Float = 0f)
